further function extraction
