+++ src/client/components/pinboard-tile/pinboard-dataset.tsx
@@ -1,0 +1,45 @@
+/*
+ * Copyright 2017-2018 Allegro.pl
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { Datum } from "plywood";
+import * as React from "react";
+import { Dimension } from "../../../common/models/dimension/dimension";
+import { Unary } from "../../../common/utils/functional/functional";
+import { DataRows } from "./utils/data-rows";
+import { EditMode } from "./utils/edit-mode";
+import filterByDimensionValue from "./utils/filter-by-dimension-value";
+
+interface PinboardDatasetProps {
+  editMode: EditMode;
+  data: Datum[];
+  searchText: string;
+  dimension: Dimension;
+  formatter: Unary<Datum, string>;
+}
+
+export const PinboardDataset: React.SFC<PinboardDatasetProps> = props => {
+  const { editMode, data, dimension, searchText, formatter } = props;
+  const filteredData = filterByDimensionValue(data, dimension, searchText);
+  const emptySearchResults = searchText && filteredData.length === 0;
+  return <div className="rows">
+    <DataRows editMode={editMode}
+              data={filteredData}
+              searchText={searchText}
+              dimension={dimension}
+              formatter={formatter} />
+    {emptySearchResults && <div className="message">{`No results for "${searchText}"`}</div>}
+  </div>;
+};
