+++ src/client/components/pinboard-tile/utils/filter-by-dimension-value.ts
@@ -1,0 +1,24 @@
+/*
+ * Copyright 2017-2018 Allegro.pl
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { Datum } from "plywood";
+import { Dimension } from "../../../../common/models/dimension/dimension";
+
+export default function filterByDimensionValue(data: Datum[], dimension: Dimension, searchText: string): Datum[] {
+  if (!searchText) return data;
+  const lowerSearchText = searchText.toLowerCase();
+  return data.filter(datum => String(datum[dimension.name]).includes(lowerSearchText));
+}
